[DR-3149] Return 404 instead of 403 if resource does not exist (#1502)
* Check if dataset exists before verifying auth

* Add missing dataset controller auth checks

- The applyDatasetDataDeletion endpoint was
  verifying authorization in the first step of
  the flight instead of in the controller like
  the other endpoints. The verify auth flight step
  was not used anywhere else.
- The bulkFileLoad and bulkFileLoadArray endpoints
  were not verifying auth in the controller, unlike
  other ingest endpoints

* Fix tests and spotbugs errors

* Return a 404 error if snaphot does not exist

* Return a 404 error if profile does not exist

* Fix tests

* Remove commented lines

+++ src/main/java/bio/terra/app/controller/DatasetsApiController.java
@@ -56,6 +56,7 @@ import bio.terra.service.job.JobService;
 import bio.terra.service.job.exception.InvalidJobParameterException;
 import io.swagger.annotations.Api;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
@@ -148,11 +149,8 @@ public class DatasetsApiController implements DatasetsApi {
               defaultValue = RETRIEVE_INCLUDE_DEFAULT_VALUE)
           List<DatasetRequestAccessIncludeModel> include) {
     AuthenticatedUserRequest userRequest = getAuthenticatedInfo();
-    iamService.verifyAuthorizations(
-        userRequest,
-        IamResourceType.DATASET,
-        id.toString(),
-        DatasetService.getRetrieveDatasetRequiredActions(include));
+    verifyDatasetAuthorizations(
+        userRequest, id.toString(), DatasetService.getRetrieveDatasetRequiredActions(include));
 
     return ResponseEntity.ok(datasetService.retrieveDatasetModel(id, userRequest, include));
   }
@@ -160,7 +158,7 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<DatasetSummaryModel> retrieveDatasetSummary(UUID id) {
     AuthenticatedUserRequest userRequest = getAuthenticatedInfo();
-    iamService.verifyAuthorization(userRequest, IamResourceType.DATASET, id.toString());
+    verifyDatasetAuthorization(userRequest, id.toString(), null);
     return ResponseEntity.ok(datasetService.retrieveDatasetSummary(id));
   }
 
@@ -174,8 +172,7 @@ public class DatasetsApiController implements DatasetsApi {
       SqlSortDirection direction,
       String filter) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.READ_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.READ_DATA);
     // TODO: Remove after https://broadworkbench.atlassian.net/browse/DR-2588 is fixed
     SqlSortDirection sortDirection = Objects.requireNonNullElse(direction, SqlSortDirection.ASC);
     DatasetDataModel previewModel =
@@ -187,8 +184,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<ColumnStatisticsModel> lookupDatasetColumnStatisticsById(
       UUID id, String table, String column, String filter) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.READ_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.READ_DATA);
     ColumnStatisticsModel columnStatisticsModel =
         datasetService.retrieveColumnStatistics(userReq, id, table, column, filter);
     return ResponseEntity.ok(columnStatisticsModel);
@@ -197,8 +193,7 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<JobModel> deleteDataset(@PathVariable("id") UUID id) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.DELETE);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.DELETE);
     String jobId = datasetService.delete(id.toString(), userReq);
     // we can retrieve the job we just created
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
@@ -209,7 +204,7 @@ public class DatasetsApiController implements DatasetsApi {
       UUID id, DatasetPatchRequestModel patchRequest) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
     Set<IamAction> actions = datasetService.patchDatasetIamActions(patchRequest);
-    iamService.verifyAuthorizations(userReq, IamResourceType.DATASET, id.toString(), actions);
+    verifyDatasetAuthorizations(userReq, id.toString(), actions);
     return ResponseEntity.ok(datasetService.patch(id, patchRequest, userReq));
   }
 
@@ -239,8 +234,7 @@ public class DatasetsApiController implements DatasetsApi {
       ingest.updateStrategy(UpdateStrategyEnum.APPEND);
     }
     validateIngestParams(ingest, id);
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = datasetService.ingestDataset(id.toString(), ingest, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -249,8 +243,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> addDatasetAssetSpecifications(
       @PathVariable("id") UUID id, @Valid @RequestBody AssetModel asset) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.MANAGE_SCHEMA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.MANAGE_SCHEMA);
     String jobId = datasetService.addDatasetAssetSpecifications(id.toString(), asset, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -259,8 +252,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> removeDatasetAssetSpecifications(
       @PathVariable("id") UUID id, @PathVariable("assetid") String assetId) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.MANAGE_SCHEMA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.MANAGE_SCHEMA);
     String jobId = datasetService.removeDatasetAssetSpecifications(id.toString(), assetId, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -269,6 +261,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> applyDatasetDataDeletion(
       UUID id, @RequestBody @Valid DataDeletionRequest dataDeletionRequest) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.SOFT_DELETE);
     String jobId = datasetService.deleteTabularData(id.toString(), dataDeletionRequest, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -278,8 +271,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> deleteFile(
       @PathVariable("id") UUID id, @PathVariable("fileid") String fileid) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.SOFT_DELETE);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.SOFT_DELETE);
     String jobId = fileService.deleteFile(id.toString(), fileid, userReq);
     // we can retrieve the job we just created
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
@@ -288,8 +280,7 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<List<FileModel>> listFiles(UUID id, Integer offset, Integer limit) {
     AuthenticatedUserRequest userRequest = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userRequest, IamResourceType.DATASET, id.toString(), IamAction.READ_DATASET);
+    verifyDatasetAuthorization(userRequest, id.toString(), IamAction.READ_DATASET);
     List<FileModel> results = fileService.listDatasetFiles(id.toString(), offset, limit);
     return ResponseEntity.ok(results);
   }
@@ -298,8 +289,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> ingestFile(
       @PathVariable("id") UUID id, @Valid @RequestBody FileLoadModel ingestFile) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = fileService.ingestFile(id.toString(), ingestFile, userReq);
     // we can retrieve the job we just created
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
@@ -309,6 +299,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> bulkFileLoad(
       @PathVariable("id") UUID id, @Valid @RequestBody BulkLoadRequestModel bulkFileLoad) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = fileService.ingestBulkFile(id.toString(), bulkFileLoad, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -318,6 +309,7 @@ public class DatasetsApiController implements DatasetsApi {
       @PathVariable("id") UUID id,
       @Valid @RequestBody BulkLoadArrayRequestModel bulkFileLoadArray) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = fileService.ingestBulkFileArray(id.toString(), bulkFileLoadArray, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -328,8 +320,7 @@ public class DatasetsApiController implements DatasetsApi {
       @PathVariable("loadTag") String loadTag,
       @Valid @RequestParam(value = "offset", required = false, defaultValue = "0") Integer offset,
       @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") Integer limit) {
-    iamService.verifyAuthorization(
-        getAuthenticatedInfo(), IamResourceType.DATASET, id.toString(), IamAction.READ_DATASET);
+    verifyDatasetAuthorization(getAuthenticatedInfo(), id.toString(), IamAction.READ_DATASET);
     List<BulkLoadHistoryModel> history = datasetService.getLoadHistory(id, loadTag, offset, limit);
     return ResponseEntity.ok(new BulkLoadHistoryModelList().total(history.size()).items(history));
   }
@@ -339,8 +330,7 @@ public class DatasetsApiController implements DatasetsApi {
       @PathVariable("id") UUID id,
       @PathVariable("fileid") String fileid,
       @RequestParam(value = "depth", required = false, defaultValue = "0") Integer depth) {
-    iamService.verifyAuthorization(
-        getAuthenticatedInfo(), IamResourceType.DATASET, id.toString(), IamAction.READ_DATA);
+    verifyDatasetAuthorization(getAuthenticatedInfo(), id.toString(), IamAction.READ_DATA);
     FileModel fileModel = fileService.lookupFile(id.toString(), fileid, depth);
     return ResponseEntity.ok(fileModel);
   }
@@ -351,8 +341,7 @@ public class DatasetsApiController implements DatasetsApi {
       @RequestParam(value = "path") String path,
       @RequestParam(value = "depth", required = false, defaultValue = "0") Integer depth) {
 
-    iamService.verifyAuthorization(
-        getAuthenticatedInfo(), IamResourceType.DATASET, id.toString(), IamAction.READ_DATA);
+    verifyDatasetAuthorization(getAuthenticatedInfo(), id.toString(), IamAction.READ_DATA);
     if (!ValidationUtils.isValidPath(path)) {
       throw new ValidationException("InvalidPath");
     }
@@ -410,8 +399,7 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<JobModel> updateSchema(UUID id, DatasetSchemaUpdateModel body) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.MANAGE_SCHEMA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.MANAGE_SCHEMA);
     String jobId = datasetService.updateDatasetSchema(id, body, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -419,8 +407,7 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<JobModel> openTransaction(UUID id, TransactionCreateModel body) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = datasetService.openTransaction(id, body, userReq);
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -429,8 +416,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<JobModel> closeTransaction(
       UUID id, UUID transactionId, TransactionCloseModel body) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.INGEST_DATA);
     String jobId = datasetService.closeTransaction(id, transactionId, userReq, body.getMode());
     return jobToResponse(jobService.retrieveJob(jobId, userReq));
   }
@@ -438,15 +424,13 @@ public class DatasetsApiController implements DatasetsApi {
   @Override
   public ResponseEntity<List<TransactionModel>> enumerateTransactions(
       UUID id, Integer offset, Integer limit) {
-    iamService.verifyAuthorization(
-        getAuthenticatedInfo(), IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(getAuthenticatedInfo(), id.toString(), IamAction.INGEST_DATA);
     return ResponseEntity.ok(datasetService.enumerateTransactions(id, offset, limit));
   }
 
   @Override
   public ResponseEntity<TransactionModel> retrieveTransaction(UUID id, UUID transactionId) {
-    iamService.verifyAuthorization(
-        getAuthenticatedInfo(), IamResourceType.DATASET, id.toString(), IamAction.INGEST_DATA);
+    verifyDatasetAuthorization(getAuthenticatedInfo(), id.toString(), IamAction.INGEST_DATA);
     return ResponseEntity.ok(datasetService.retrieveTransaction(id, transactionId));
   }
 
@@ -454,8 +438,7 @@ public class DatasetsApiController implements DatasetsApi {
   public ResponseEntity<DatasetSummaryModel> updateDatasetTags(
       UUID id, TagUpdateRequestModel tagUpdateRequest) {
     AuthenticatedUserRequest userReq = getAuthenticatedInfo();
-    iamService.verifyAuthorization(
-        userReq, IamResourceType.DATASET, id.toString(), IamAction.MANAGE_SCHEMA);
+    verifyDatasetAuthorization(userReq, id.toString(), IamAction.MANAGE_SCHEMA);
     return ResponseEntity.ok(datasetService.updateTags(id, tagUpdateRequest));
   }
 
@@ -509,5 +492,26 @@ public class DatasetsApiController implements DatasetsApi {
     if (!errors.isEmpty()) {
       throw new InvalidJobParameterException("Invalid ingest parameters detected", errors);
     }
+  }
+
+  private void verifyDatasetAuthorization(
+      AuthenticatedUserRequest userReq, String resourceId, IamAction action) {
+    IamResourceType resourceType = IamResourceType.DATASET;
+    // Check if dataset exists
+    datasetService.retrieveDatasetSummary(UUID.fromString(resourceId));
+    // Verify dataset permissions
+    if (action != null) {
+      iamService.verifyAuthorization(userReq, resourceType, resourceId, action);
+    } else {
+      iamService.verifyAuthorization(userReq, resourceType, resourceId);
+    }
+  }
+
+  private void verifyDatasetAuthorizations(
+      AuthenticatedUserRequest userReq, String resourceId, Collection<IamAction> actions) {
+    // Check if dataset exists
+    datasetService.retrieveDatasetSummary(UUID.fromString(resourceId));
+    // Verify dataset permissions
+    iamService.verifyAuthorizations(userReq, IamResourceType.DATASET, resourceId, actions);
   }
 }
